refactor: move snapshot tests to Mockolate.Tests project
