+++ Tests/Mockolate.Tests/MockCreationTests.cs
@@ -1,0 +1,7 @@
+#if NET10_0_OR_GREATER
+using System.Net.Http;
+using Mockolate.Tests.GeneratorCoverage;
+
+namespace Mockolate.Tests;
+
+#endif
